requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.0.0
matplotlib>=3.7.0
fpdf2>=2.7.0
openpyxl>=3.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.12.0
google-generativeai>=0.5.0
kaleido>=0.2.1
